Consolidate these screens into one in common.
