+++ lib/l10n/app_en.arb
@@ -530,9 +530,7 @@
   "proposalVoteConfirmVote": "Vote",
   "proposalWeightedVoteConfirmWeightedVote": "Confirm Weighted Vote",
   "proposalWeightedVoteProposalId": "Proposal ID#",
-  "proposalVoteSuccessSuccess": "SUCCESS",
-  "proposalVoteSuccessVoteSuccessful": "Vote Successful",
-  "proposalVoteSuccessBackToDashboard": "Back To Dashboard",
+  "proposalVoteComplete": "VOTE COMPLETE",
   "stakingConfirmUndelegationDetails": "Undelegation Details",
   "stakingConfirmAmountToUndelegate": "Amount to Undelegate",
   "stakingConfirmNewTotalDelegation": "New Total Delegation",
