+++ lib/screens/home/proposals/proposals_flow.dart
@@ -2,7 +2,7 @@ import 'package:provenance_dart/proto_gov.dart' as proto;
 import 'package:provenance_wallet/common/flow_base.dart';
 import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/pw_data_screen.dart';
-import 'package:provenance_wallet/screens/home/proposals/proposal_success/proposal_success_screen.dart';
+import 'package:provenance_wallet/common/widgets/pw_transaction_complete_screen.dart';
 import 'package:provenance_wallet/screens/home/proposals/proposal_vote_confirm/proposal_vote_confirm_screen.dart';
 import 'package:provenance_wallet/screens/home/proposals/proposal_weighted_vote/proposal_weighted_vote_screen.dart';
 import 'package:provenance_wallet/screens/home/proposals/proposal_weighted_vote_confirm/proposal_weighted_vote_confirm.dart';
@@ -14,6 +14,7 @@ import 'package:provenance_wallet/services/account_service/account_service.dart'
 import 'package:provenance_wallet/services/models/account.dart';
 import 'package:provenance_wallet/services/models/proposal.dart';
 import 'package:provenance_wallet/util/get.dart';
+import 'package:provenance_wallet/util/strings.dart';
 
 abstract class ProposalsFlowNavigator {
   Future<void> showProposalDetails(
@@ -38,7 +39,7 @@ abstract class ProposalsFlowNavigator {
     String screenTitle,
   );
 
-  Future<void> showTransactionSuccess();
+  Future<void> showTransactionComplete(Object? response);
 
   void onComplete();
 
@@ -123,9 +124,15 @@ class _ProposalsFlowState extends FlowBaseState<ProposalsFlow>
   }
 
   @override
-  Future<void> showTransactionSuccess() async {
+  Future<void> showTransactionComplete(Object? response) async {
     showPage(
-      (context) => ProposalSuccessScreen(),
+      (context) => PwTransactionCompleteScreen(
+        title: Strings.of(context).proposalVoteComplete,
+        onBackToDashboard: backToDashboard,
+        response: response,
+        onComplete: onComplete,
+        onPressed: () => showTransactionData(response, "Response"),
+      ),
     );
   }
